Added explicit casting to clear warnings
git-svn-id: svn://localhost/media/c/Documnts/src/flare/repo.svndump/trunk@167 6ae5efe3-03d3-4ed7-8a64-3d71375a9a67

+++ src/PowerManager.cpp
@@ -554,7 +554,7 @@ bool PowerManager::groundRay(int power_index, StatBlock *src_stats, Point target
 		location_iterator.y += speed.y;
 
 		// only travels until it hits a wall
-		if (!collider->is_empty(location_iterator.x, location_iterator.y)) {
+		if (!collider->is_empty((int)location_iterator.x, (int)location_iterator.y)) {
 			break; // no more hazards
 		}
